+++ NEWS.md
@@ -1,4 +1,5 @@
 # SAVM (devel)
 
-* `compute_fetch()` has a new argument `n_bearings` that provides the number of bearings, it replaces `n_quad_seg` (see #4). 
+* `compute_fetch()` has a new argument `n_bearings` that provides the number of bearings, it replaces `n_quad_seg` (see #4 and #5). 
+* `compute_fetch()` only compute the mean fetch for all bearings (see #4). 
 * `invert_polygon()` has been added to invert polygon (see #6)